backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
PyYAML==6.0
django-cors-headers==3.13.0
gunicorn==20.1.0
python-dotenv==1.0.0